Refactor project structure and update import paths for modules
