+++ .github/workflows/ci.yml
@@ -77,7 +77,7 @@ jobs:
         run: |
           cd ${{ matrix.service }}
           uv pip install -e .
-          uvx pytest@latest --junitxml=report.xml
+          uvx pytest tests/ --junitxml=report.xml
         continue-on-error: false
 
       - name: Upload coverage report
