+++ services/convertor/pyproject.toml
@@ -57,3 +57,6 @@ exclude = "(alembic|.venv)"
 dev = [
     "pytest-cov>=6.2.1",
 ]
+
+[tool.setuptools.packages.find]
+where = ["convertor"]
