+++ services/convertor/tests/test_s3_client.py
@@ -1,4 +1,4 @@
-from ..s3_client import s3_client
+from convertor.s3_client import s3_client
 
 
 def test_import_s3_client() -> None:
